refact: moves comment query and data request status queries to their own repository
